event publisher도입, 헬퍼 컴포넌트 도입
